Resolve issue #13
When the application of a time offset would result in a negative begin
or end time, rather than terminating with an error, prune the p or span
element that would have that attribute, and issue a warning.

Also includes some tidying of duplicated code where begin and end
attributes can be handled with a single template.